FeedBack para las dinamicas (quiz, columnas, puzzle y codigo) ademas de las rutas correctas a las images
