+++ lib/features/quiz_game/view/widgets/quiz_page.dart
@@ -5,11 +5,24 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:kitsucode/features/quiz_game/view/widgets/result_page.dart';
+// import 'package:kitsucode/features/quiz_game/view/widgets/result_page.dart'; // <-- ELIMINADO
 import 'package:kitsucode/core/utils/app_colors.dart';
 import 'package:kitsucode/features/quiz_game/view/quiz_loader.dart'; // Importa QuizData
 
-class QuizPage extends StatefulWidget {
+// --- ¡NUEVOS IMPORTS! (Copiados de result_page.dart) ---
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:kitsucode/features/challenge/repository/challenge_repository.dart';
+import 'package:kitsucode/shared/appbar/app_bar_provider.dart';
+import 'package:kitsucode/features/competences/provider/ranking_provider.dart';
+import 'package:go_router/go_router.dart';
+import 'package:kitsucode/features/challenge/widgets/challenge_feedback_modal.dart';
+import 'package:kitsucode/core/utils/app_themes.dart';
+import 'package:kitsucode/features/challenge/view/feedback/challenge_failure_view.dart' show RecursoModel;
+// --- FIN NUEVOS IMPORTS ---
+
+
+// --- ¡Convertido a ConsumerStatefulWidget! ---
+class QuizPage extends ConsumerStatefulWidget {
   final QuizData mydata;
   final String retoId;
 
@@ -20,11 +33,11 @@ class QuizPage extends StatefulWidget {
   });
   
   @override
-  State<QuizPage> createState() => _QuizPageState();
+  ConsumerState<QuizPage> createState() => _QuizPageState();
 }
 
-class _QuizPageState extends State<QuizPage> {
-  // ... (Tus variables de estado: marks, i, timer, etc. no cambian) ...
+// --- ¡Convertido a ConsumerState! ---
+class _QuizPageState extends ConsumerState<QuizPage> { 
   int marks = 0;
   int i = 0;
   bool disableAnswer = false;
@@ -35,10 +48,10 @@ class _QuizPageState extends State<QuizPage> {
   int totalQuestions = 0;
   String? selectedAnswer;
   bool _cancelTimer = false;
+  bool _hasSubmitted = false; 
 
   @override
   void initState() {
-    // ... (Tu initState no cambia) ...
     super.initState();
     _startTimer();
     _genRandomArray();
@@ -49,13 +62,11 @@ class _QuizPageState extends State<QuizPage> {
 
   @override
   void dispose() {
-    // ... (Tu dispose no cambia) ...
     _cancelTimer = true;
     super.dispose();
   }
 
   void _genRandomArray() {
-    // ... (Tu función _genRandomArray no cambia) ...
     if (widget.mydata.questions.isNotEmpty) {
       totalQuestions = widget.mydata.totalQuestions;
       var rand = Random();
@@ -72,7 +83,6 @@ class _QuizPageState extends State<QuizPage> {
   }
 
   void _startTimer() {
-    // ... (Tu función _startTimer no cambia) ...
     const onesec = Duration(seconds: 1);
     Timer.periodic(onesec, (Timer t) {
       if (!mounted) {
@@ -114,19 +124,14 @@ class _QuizPageState extends State<QuizPage> {
             int duration = (30 * totalQuestions) - timer;
             if (duration < 0) duration = 0; 
 
-            // --- ¡CAMBIO 2! (Pasamos el retoId y los RECURSOS) ---
-            Navigator.of(context).pushReplacement(
-              MaterialPageRoute(
-                builder: (context) => QuizResultPage(
-                  marks: marks,
-                  totalQuestions: totalQuestions,
-                  durationInSeconds: duration,
-                  retoId: widget.retoId,
-                  recursos: widget.mydata.recursos, // <-- ¡AÑADIDO!
-                ),
-              ),
+            final double scoreRatio = marks / (totalQuestions * 5);
+            final int percentage = (scoreRatio * 100).round();
+
+            _showFeedbackModal(
+              percentage: percentage,
+              durationInSeconds: duration,
+              recursos: widget.mydata.recursos,
             );
-            // --- FIN CAMBIO 2 ---
           }
           return;
         }
@@ -137,12 +142,12 @@ class _QuizPageState extends State<QuizPage> {
     _startTimer();
   }
 
+  // --- ¡¡¡AQUÍ ESTÁ LA CORRECCIÓN DEL BUG DE PUNTUACIÓN!!! ---
   void _checkAnswer(String k, ColorScheme pythonColorScheme) {
-    // ... (Tu función _checkAnswer no cambia) ...
     String questionKey = widget.mydata.questions.keys.elementAt(i);
-    if (k.isNotEmpty &&
-        widget.mydata.answers[questionKey] ==
-            widget.mydata.options[questionKey]![k]) {
+    
+    // Compara la 'letra' seleccionada (k) con la 'letra' de la respuesta (answers[questionKey])
+    if (k.isNotEmpty && widget.mydata.answers[questionKey] == k) {
       marks = marks + 5;
     }
 
@@ -153,8 +158,94 @@ class _QuizPageState extends State<QuizPage> {
       });
     }
   }
+  // --- FIN DE LA CORRECCIÓN ---
+
+
+  // --- Lógica del modal (movida de result_page) ---
+  ThemeData _getLanguageTheme(String langName, Brightness brightness) {
+    final isDark = brightness == Brightness.dark;
+    
+    switch (langName.toLowerCase().trim()) {
+      case 'python':
+        return isDark ? AppThemes.pythonDarkTheme : AppThemes.pythonTheme;
+      case 'c':
+        return isDark ? AppThemes.cDarkTheme : AppThemes.cTheme;
+      case 'java':
+        return isDark ? AppThemes.javaDarkTheme : AppThemes.javaTheme;
+      default:
+        return isDark ? AppThemes.darkTheme : AppThemes.lightTheme;
+    }
+  }
+
+  void _showFeedbackModal({
+    required int percentage,
+    required int durationInSeconds,
+    required List<RecursoModel> recursos,
+  }) {
+    if (_hasSubmitted) return; 
+    
+    final bool esCorrecto = (percentage > 50);
+
+    final appBarState = ref.read(appBarProvider);
+    final challengeTheme = _getLanguageTheme(
+      appBarState.languageName,
+      Theme.of(context).brightness,
+    );
+
+    showModalBottomSheet(
+      context: context,
+      backgroundColor: Colors.transparent,
+      isScrollControlled: true,
+      useRootNavigator: true, 
+      builder: (ctx) { 
+        return Theme(
+          data: challengeTheme,
+          child: ChallengeFeedbackModal(
+            isCorrect: esCorrecto,
+            onContinue: () async {
+              
+              Navigator.of(ctx).pop(); 
+              
+              if (_hasSubmitted) return;
+              _hasSubmitted = true; 
+
+              final repository = ref.read(challengeRepositoryProvider);
+              final int retoIdAsInt = int.parse(widget.retoId);
+
+              if (esCorrecto) {
+                final int trofeos = await repository.submitChallengeAttempt(
+                  retoId: retoIdAsInt,
+                  fueExitoso: true,
+                  tiempoQueTardo: durationInSeconds, 
+                );
+                
+                ref.read(appBarProvider.notifier).fetchStats();
+                ref.invalidate(globalRankingProvider);
+                
+                if (!context.mounted) return;
+                context.pushReplacement('/challenge_success', extra: trofeos);
+              
+              } else {
+                 await repository.submitChallengeAttempt(
+                  retoId: retoIdAsInt,
+                  fueExitoso: false,
+                  tiempoQueTardo: durationInSeconds,
+                );
+
+                ref.read(appBarProvider.notifier).fetchStats();
+                
+                if (!context.mounted) return;
+                context.pushReplacement('/challenge_failure', extra: recursos);
+              }
+            },
+          ),
+        );
+      },
+    );
+  }
+  
   // ... (El resto de tu código: _choiceButton, build, _buildDuolingoQuestionArea...
-  // ... no necesitan cambios) ...
+  // ... no necesitan cambios y van aquí) ...
   
   Widget _choiceButton(String k, ColorScheme pythonColorScheme) {
     String questionKey = widget.mydata.questions.keys.elementAt(i);
@@ -166,11 +257,13 @@ class _QuizPageState extends State<QuizPage> {
 
     if (disableAnswer) {
       String correctAnswerKey = '';
-      widget.mydata.options[questionKey]!.forEach((key, value) {
-        if (value == widget.mydata.answers[questionKey]) {
-          correctAnswerKey = key;
-        }
-      });
+      // ¡OJO! Aquí estaba la respuesta correcta
+      final String respuestaCorrectaLetra = widget.mydata.answers[questionKey]!;
+      
+      // Buscamos la 'key' ('a', 'b', 'c', 'd') que coincide con la letra de la respuesta
+      // (En tu JSON, la 'respuesta' YA ES la 'key', así que esto es directo)
+      correctAnswerKey = respuestaCorrectaLetra;
+
 
       if (k == correctAnswerKey) {
         buttonColor = Colors.green.withAlpha(51);
@@ -290,7 +383,7 @@ class _QuizPageState extends State<QuizPage> {
                       TextButton(
                         onPressed: () {
                           Navigator.of(context).pop();
-                          Navigator.of(context).pop();
+                          context.pop(); 
                         },
                         child: const Text('Salir'),
                       ),
@@ -391,6 +484,7 @@ class _QuizPageState extends State<QuizPage> {
                 child: ElevatedButton(
                   style: ElevatedButton.styleFrom(
                     backgroundColor: disableAnswer
+                        // AHORA ESTO FUNCIONARÁ
                         ? (marks > (j - 1) * 5 ? Colors.green : Colors.red)
                         : (selectedAnswer != null
                             ? pythonColorScheme
