+++ pubspec.yaml
@@ -41,6 +41,7 @@ flutter:
     - assets/images/
     - assets/images/navbar/
     - assets/images/home/
+    - assets/images/auth/
     #- assets/icons/
     #- assets/fonts/
     - assets/.env
